+++ packages/transaction-controller/src/utils/feature-flags.test.ts
@@ -74,7 +74,7 @@ describe('Feature Flags Utils', () => {
     getFeatureFlagsMock.mockReturnValue({
       cacheTimestamp: 0,
       remoteFeatureFlags: featureFlags,
-      rawProcessedRemoteFeatureFlags: {},
+      rawRemoteFeatureFlags: {},
       localOverrides: {},
     });
   }